style

+++ inc/paramdepend.h
@@ -727,7 +727,7 @@ extern "C" {
 
 #elif (__GIC_PRESENT == 1)
 
-	typedef uint8_t IRQL_t;
+	typedef uint32_t IRQL_t;
 
 	enum
 	{
